migrations et seeders et erreur

+++ chevaucheursDeVers/database/migrations/2024_03_11_073728_create_partie_table.php
@@ -20,11 +20,11 @@ class CreatePartieTable extends Migration
             $table->boolean("partie_privee")->default(false);
             $table->boolean("est_commencee")->default(false);
             $table->boolean("est_terminee")->default(false);
-            $table->unsignedBigInteger("id_user_gagnant")->nullable();
+            $table->integer("id_user_gagnant")->nullable();
             $table->foreign("id_user_gagnant")->references("id")->on("users")->onDelete("cascade");
             $table->integer("nombre_joueurs")->default(2);
             $table->time("temps_par_coup")->default('00:01:00');
-            $table->unsignedBigInteger("id_user_host");
+            $table->integer("id_user_host");
             $table->foreign("id_user_host")->references("id")->on("users")->onDelete("cascade");
         });
     }
